+++ src/it/resources/transactionTest/purchaseContruct/bodies.txt
