+++ src/it/resources/transactionTest/purchaseContruct/contractTrees.txt
