+++ src/it/resources/transactionTest/purchaseContruct/evmCode.txt
