+++ src/it/resources/transactionTest/purchaseContruct/headers.txt
