+++ src/it/resources/transactionTest/purchaseContruct/receipts.txt
