+++ src/it/resources/transactionTest/purchaseContruct/stateTree.txt
