+++ src/it/scala/io/iohk/ethereum/transactionTest/util/DumpChainActor.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.evm.util
+package io.iohk.ethereum.transactionTest.util
 
 import java.io.FileWriter
