+++ src/it/scala/io/iohk/ethereum/transactionTest/util/DumpChainApp.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.evm.util
+package io.iohk.ethereum.transactionTest.util
 
 import akka.actor.ActorSystem
 import akka.agent.Agent
